requirements.txt
flashtorch==0.1.3
imageio==2.31.1
ipdb==0.13.13
matplotlib==3.7.1
numpy==1.22.3
opencv_contrib_python==4.6.0.66
opencv_python==4.6.0.66
opencv_python_headless==4.10.0.82
Pillow==10.4.0
thop==0.1.1.post2209072238
timm==0.9.10
torch==2.3.0
torchcam==0.4.0
torchstat==0.0.7
torchvision==0.18.0+cu121
tqdm==4.65.0
